Add support for games without `default`
This adds support for games like Mineclonia, and any game xcompat adds
support for. The only asterix is dye support, the pattern for dyes has
been expanded for common dye mods, but if a new dye mod has different
dye names those might not work without an update to the pattern.